Refactor for better separation
