+++ src/index.ts
@@ -1,5 +1,5 @@
 import SyncDb from './SyncDb';
-import { synchronize } from './migrator';
+import { synchronize } from './services/syncer';
 import { loadConfig, resolveConnections } from './config';
 
 export { SyncDb, loadConfig, synchronize, resolveConnections };
