+++ src/migrator.ts
@@ -1,17 +1,9 @@
 import * as Knex from 'knex';
-import { mergeDeepRight } from 'ramda';
 
-import Connection from './Connection';
+import { dbLogger } from './logger';
 import * as sqlRunner from './sqlRunner';
-import { log, dbLogger } from './logger';
-import { getConnectionId } from './config';
-import SyncParams from './domain/SyncParams';
-import SyncConfig from './domain/SyncConfig';
 import SyncResult from './domain/SyncResult';
 import SyncContext from './domain/SyncContext';
-import { DEFAULT_SYNC_PARAMS } from './constants';
-import ConnectionConfig from './domain/ConnectionConfig';
-
 import * as configInjection from './services/configInjection';
 
 /**
@@ -86,69 +78,13 @@ async function teardown(trx: Knex.Transaction, context: SyncContext): Promise<vo
 }
 
 /**
- * Synchronize all the configured database connections.
- *
- * @param {SyncConfig} config
- * @param {(ConnectionConfig[] | ConnectionConfig | Knex[] | Knex)} conn
- * @param {SyncParams} [options]
- * @returns {Promise<SyncResult[]>}
- */
-export async function synchronize(
-  config: SyncConfig,
-  conn: ConnectionConfig[] | ConnectionConfig | Knex[] | Knex,
-  options?: SyncParams
-): Promise<SyncResult[]> {
-  log('Starting to synchronize.');
-  const connArr = Array.isArray(conn) ? conn : [conn];
-  const connections = mapToConnectionInstances(connArr);
-  const params = mergeDeepRight(DEFAULT_SYNC_PARAMS, options);
-  const cliEnvironment = process.env.SYNC_DB_CLI === 'true';
-  const promises = connections.map(connection =>
-    synchronizeDatabase(connection.getInstance(), {
-      config,
-      params,
-      cliEnvironment,
-      connectionId: getConnectionId(connection.getConfig())
-    })
-  );
-
-  const result = await Promise.all(promises);
-
-  log('Finished all');
-
-  return result;
-}
-
-/**
- * Map connection configuration list to the connection instances.
- *
- * @param {((ConnectionConfig | Knex)[])} connectionList
- * @returns {Connection[]}
- */
-function mapToConnectionInstances(connectionList: (ConnectionConfig | Knex)[]): Connection[] {
-  return connectionList.map(con => {
-    // TODO: Ask for id in for programmatic API too -
-    // when Knex instance is passed directly.
-    if (Connection.isKnexInstance(con)) {
-      log(`Received connection instance to database: ${con.client.config.connection.database}`);
-
-      return Connection.withKnex(con);
-    }
-
-    log(`Received connection config to database: ${con.database}`);
-
-    return new Connection(con);
-  });
-}
-
-/**
  * Synchronize on a single database connection.
  *
  * @param {Knex} connection
  * @param {SyncContext} context
  * @returns {Promise<SyncResult>}
  */
-async function synchronizeDatabase(connection: Knex, context: SyncContext): Promise<SyncResult> {
+export async function synchronizeDatabase(connection: Knex, context: SyncContext): Promise<SyncResult> {
   const { connectionId } = context;
   const logDb = dbLogger(connectionId);
   const result: SyncResult = { connectionId, success: false };
